Refactoring part 2

This is a mix of trivial refactoring changes (2c8f82f, 474cb72, a1b33e0), C++ conformance fixes (9202216) and a change to how an important assumption is validated inside the code (05a297c).

Note that Windows support is dependent on these changes, so it needs to be merged before the Windows support is.

+++ include/sharpSAT/component_types/component.h
@@ -9,6 +9,7 @@
 #define SHARP_SAT_COMPONENT_H_
 
 #include <assert.h>
+#include <cstddef>
 #include <vector>
 
 #include <sharpSAT/primitive_types.h>
@@ -66,7 +67,7 @@ public:
     return clauses_ofs_ - 1;
   }
 
-  unsigned numLongClauses() const {
+  size_t numLongClauses() const {
     return data_.size() - clauses_ofs_ - 1;
   }
